test(cli): split serve tests
